+++ app.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import StarwarsQuote from './components/star-wars-quote';
+import StarwarsQuote from './index.js';
 
 const rootElement = document.getElementById('container');
